\name{sum_square-package}
\alias{sum_square-package}
\alias{sum_square}
\docType{package}
\title{
\packageTitle{sum_square}
}
\description{
\packageDescription{sum_square}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sum_square}
\packageIndices{sum_square}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sum_square}

Maintainer: \packageMaintainer{sum_square}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
