% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_square-data.R, R/sum_square-package.R
\docType{data}
\name{sum_square}
\alias{sum_square}
\alias{sum_sq}
\alias{sum_square}
\alias{sum_sq}
\alias{sum_square-package}
\title{Help page title. Probably the package name and tagline.}
\description{
A description of what the package does, why you might want to use it,
which functions to look at first, and anything else that the user
really, absolutely, must look at because you've created it and it is
astonishing.

A description of what the package does, why you might want to use it,
which functions to look at first, and anything else that the user
really, absolutely, must look at because you've created it and it is
astonishing.
}
\author{
QH

QH
}
\keyword{dataset}
\keyword{package}
